[TASK] Migrate AbstractImporter into XmlImporter
The importer can only import XML files. Sure, technically you also
can import event records of any other file type, but the PHP class
must consist in events2 itself. That is not really extensible and that is
the reason why we have removed the abstract and the interface.

We have cleaned up the Task\Importer a little bit and wrapped
GU::makeInstance into own getter methods.